refactor: move into folder and use immer reducer
