+++ packages/frontend/package.json
@@ -60,7 +60,8 @@
     "regenerator-runtime": "^0.13.5",
     "reselect": "^4.0.0",
     "styled-components": "^5.3.0",
-    "timeago.js": "^4.0.2"
+    "timeago.js": "^4.0.2",
+    "use-immer": "^0.7.0"
   },
   "scripts": {
     "predeploy": "yarn build",
